major overhaul. Optimized code, improved parsing.
Changes: 1. Separated parsing code from display code. 2. Program now has built in stricter checking for correctly formatting, so terms should be strictly separated by a white space (ascii 32) a dash (ascii 45) and another space. If it finds characters that are similar, it will replace them on the fly so that the terms and definitions or the flashcards are still loaded. Also, removes initial numbers from the file, so on a line such as: 
1. hi - bonjour 
the number is removed. Added some common numbering schemes. Also replaces non breaking space andother invisible stuff that might get copied from a website, or a pdf, ETC.
3. Refuses to load severely malformed files, announces which lines have errors. Also, a separate debug window is available that shows where parsing went wrong, its accessed by pressing the cliché f12.